requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.2.7
django-nested-admin==3.3.3
django-solo==1.1.5
filelock==3.0.12
idna==2.10
pipenv==2020.8.13
pycosat==0.6.3
python-monkey-business==1.0.0
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.2.7
django-environ==0.7.0
django-nested-admin==3.3.3
django-solo==1.1.5
filelock==3.0.12
idna==2.10
pipenv==2020.8.13
pycosat==0.6.3
python-monkey-business==1.0.0
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.2.7
django-environ==0.7.0
django-nested-admin==3.3.3
django-solo==1.1.5
filelock==3.0.12
flake8==3.9.2
idna==2.10
mccabe==0.6.1
Pillow==8.3.2
pipenv==2020.8.13
pycodestyle==2.7.0
pycosat==0.6.3
pyflakes==2.3.1
python-monkey-business==1.0.0
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.2.7
django-appconf==1.0.4
django-environ==0.7.0
django-image-cropping==1.6.1
django-nested-admin==3.3.3
django-solo==1.1.5
filelock==3.0.12
flake8==3.9.2
idna==2.10
mccabe==0.6.1
Pillow==8.3.2
pipenv==2020.8.13
pycodestyle==2.7.0
pycosat==0.6.3
pyflakes==2.3.1
python-monkey-business==1.0.0
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.2.7
django-appconf==1.0.4
django-environ==0.7.0
django-nested-admin==3.3.3
django-solo==1.1.5
filelock==3.0.12
flake8==3.9.2
idna==2.10
mccabe==0.6.1
Pillow==8.3.2
pipenv==2020.8.13
pycodestyle==2.7.0
pycosat==0.6.3
pyflakes==2.3.1
python-monkey-business==1.0.0
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.2.7
django-appconf==1.0.4
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-image-cropping==1.6.1
django-nested-admin==3.3.3
django-solo==1.1.5
filelock==3.0.12
flake8==3.9.2
idna==2.10
mccabe==0.6.1
Pillow==8.3.2
pipenv==2020.8.13
pycodestyle==2.7.0
pycosat==0.6.3
pyflakes==2.3.1
python-monkey-business==1.0.0
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
